Generate declarative config for app, proj, cluster
